model/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pillow==10.1.0
numpy==1.24.3
imageio==2.33.0
torch==2.1.0
trellisnet
gradio_litmodel3d